build: remove prescripts
pnpm does not support package.json#script prescripts such as `prestart`
and `prebuild` so make the `clean` script part of the `start` and
`build` scripts

Add rimraf as a dependency. The Netlify build environment has `NODE_ENV`
set to `production` which means that only dependencies are installed and
devDependencies are not installed.